backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
defusedxml==0.7.1
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
fonttools==4.57.0
fpdf2==2.8.3
frozenlist==1.7.0
idna==3.10
kombu==5.5.4
multidict==6.6.3
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
pypdf==5.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reportlab==4.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
twilio==9.6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
